requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
comtypes==1.2.1
cookiecutter==2.6.0
coverage==6.3.1
cryptography==45.0.3
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
distutils-pytest==0.2.1
dj-database-url==1.3.0
Django==5.1.3
django-admin-list-filter-dropdown==1.0.3
django-adminfilters==2.1.0
django-debug-toolbar==4.4.6
django-extensions==3.2.1
django-health-check==3.17.0
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.1.1
env==0.1.0
environs==9.5.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.17.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flet==0.22.1
flet-core==0.22.1
flet-runtime==0.22.1
frozenlist==1.4.1
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.10
iniconfig==1.1.1
ipykernel==6.29.0
ipython==8.21.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.6
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
openai==0.28.0
orjson==3.10.4
packaging==24.1
pandas==2.2.3
paramiko==3.5.1
parso==0.8.3
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.0.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pygame==2.5.2
Pygments==2.17.2
pymongo==3.13.0
PyNaCl==1.5.0
pyparsing==3.0.7
pypiwin32==223
pypng==0.20220715.0
pytest==7.3.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-multipart==0.0.9
python-slugify==8.0.4
pyttsx3==2.90
pytz==2024.2
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.44
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.37.2
tenacity==9.0.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.1
views==0.3
virtualenv==20.29.2
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
xlsxwriter==3.2.5
yarl==1.9.4
zeep==4.3.1
